Added code reuse requirement; see Issue #36
Added a paragraph to require agencies to actually comprise some amount (I used the 20 percent figure for now) of existing reusable code.  Personally I think this number should increase over time.  Yes, this drives the total OSS percentage of the project to be 40 percent.  I allowed an "out" for an agency to "audit" and "certify" that no reusable code was available, but I consider this unlikely except maybe on the smallest of projects.  See further discussion in Issue #36 (Incentive or requirement for actual reuse of existing code).