Update introduction.md
Added this sentence:

Further, all firms participating in or competing for future maintenance or enhancement are automatically fully informed of the code.

In my experience, this is critically important.  I expect this to be controversial, so I will create an issue to facilitate discussion.  I don't really expect this sentence to be accepted, but I thought this more concrete than a discussion in issues.