Added references to Section 508 compliance
Section 508 compliance and consideration of the IT Accessibility Standards are required as is Privacy and Security and should be considered as agency business needs. Beyond that pragmatic approach, it is required by Federal Acquisition Regulation (FAR) with technical and functional criteria as set forth by the U.S. Access Board.